Made it so we can search Workshop IDs
Basically it goes like this:
if you type the Workshop ID of an addon in the search bar, then it will list the right addon if it's installed.
If you type the Workshop ID of a collection, then it will list all of the installed addons that are in the collection.